Refactor frontend into screen-based modules
