+++ frontend/src/screens/dashboard/components/AddDataPanel.tsx
